requirements.txt
aiogram==3.10.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
pdfplumber==0.11.4
pandas==2.2.2
rapidfuzz==3.9.7
playwright==1.46.0
tqdm==4.66.4
